move and rename method page
